@@ -30,11 +30,11 @@
   "devDependencies": {
     "babel-core": "^4.7.4",
     "babel-loader": "^4.1.0",
-    "react-hot-loader": "^1.2.0",
+    "react-hot-loader": "^1.2.2",
     "webpack": "^1.7.2",
     "webpack-dev-server": "^1.7.0"
   },
   "dependencies": {
-    "react": "^0.12.2"
+    "react": "^0.13.0"
   }
 }
