@@ -1,11 +1,9 @@
 import React from 'react';
 
-const App = React.createClass({
+export default class App extends React.Component {
   render() {
     return (
       <h1>Hello, world.</h1>
     );
   }
-});
-
-export default App;
+}
